proposed/requirements.txt
tweet-preprocessor
scikit-learn
torch
transformers
wordninja